mvp auto export
